# openapi/openapi.yaml
openapi: 3.0.0
info:
    title: Ethereum Crawler API
    version: 1.0.0
    contact:
        email: amin.saradar@yahoo.com
    description: >-
      Explore the Ethereum blockchain with our powerful Ethereum Crawler.
      This tool allows you to get detailed information about any Ethereum
      fee balances for every account.

paths:
    "/status":
        get:
            operationId: GetStatus
            summary: Status
            description: |
                Returns an object containing the status response of the API.
                Meaning of status:
              
#                Address -  Ethereum address associated with the account.
#
#                TotalPaidFee - Total fees paid by the account.
#
#                LastHeight - The last known block associated with the account.
#
#                TxIndex - The last known transaction index related LastHeight.
#
#                FirstHeight - The initial block height where the account was first encountered.

            responses:
                "200":
                    $ref: '#/components/responses/StatusResponse'

    "/totalPaidFee/{address}":
        get:
            operationId: GetTotalPaidFee
            summary: Total Paid Fee
            description: Returns the total fees paid by the Ethereum account
              associated with the provided address.
            parameters:
              - in: path
                name: address
                description: the account address
                required: true
                schema:
                  type: string
                example: '0xA1E4380A3B1f749673E270229993eE55F35663b4'
            responses:
              "200":
                $ref: '#/components/responses/TotalPaidFeeResponse'
    "/firstTransaction":
      get:
        operationId: GetFirstTransaction
        summary: First Transaction
        description: |
          Returns an object containing the first transaction of the Ethereum
        responses:
          "200":
            $ref: '#/components/responses/FirstTransactionResponse'
components:
    responses:
        StatusResponse:
            description: Returns the status of crawler
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Status'
        TotalPaidFeeResponse:
          description: object  containing total paid fee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        FirstTransactionResponse:
            description: Object containing the first transaction of the Ethereum
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Status'


    schemas:


        Status:
            type: object
            required:
                - address
                - totalPaidFee
                - lastHeight
                - txIndex
                - firstHeight
            properties:
              address:
                type: string
                description: Ethereum address associated with the account.
              totalPaidFee:
                type: integer
                format: uint64
                description: Total fees paid by the account, measured in uint64.
              lastHeight:
                type: integer
                format: int64
                description: The last known transaction index related LastHeight.
              txIndex:
                type: integer
                description: Transaction index related to the account.
              firstHeight:
                type: integer
                format: int64
                description: The initial block height where the account was first encountered.